requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
influxdb==5.3.2
msgpack==1.1.2
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
urllib3==2.6.3
